+++ src/UI/ledger/LedgerLoginContainer/ConfirmAddress/ConfirmAddress.tsx
@@ -5,6 +5,7 @@ import { DataTestIdsEnum } from 'constants/index';
 import { useGetAccountInfo } from 'hooks';
 import { useSelector } from 'reduxStore/DappProviderContext';
 import { tokenLoginSelector } from 'reduxStore/selectors';
+import { getLedgerVersionOptions } from 'utils';
 
 import { WithClassnameType } from '../../../types';
 
@@ -47,6 +48,10 @@ export const ConfirmAddress = ({
     version: ledgerAccount?.version
   });
 
+  const { ledgerWithUsernames } = getLedgerVersionOptions(
+    ledgerAccount?.version ?? '0'
+  );
+
   return (
     <div
       className={classNames(
@@ -74,7 +79,9 @@ export const ConfirmAddress = ({
             ledgerModalConfirmDescriptionClassName
           )}
         >
-          For security, please confirm your address:
+          {ledgerWithUsernames
+            ? 'For security, please confirm your address:'
+            : 'For security, please confirm that your address:'}
         </div>
 
         <div
@@ -95,7 +102,7 @@ export const ConfirmAddress = ({
               ledgerModalConfirmDescriptionClassName
             )}
           >
-            and authorize:
+            {ledgerWithUsernames ? 'and authorize:' : 'and Auth Token:'}
           </div>
 
           <div
@@ -106,6 +113,16 @@ export const ConfirmAddress = ({
           >
             {authTokenText?.data}
           </div>
+          {!ledgerWithUsernames ? (
+            <div
+              className={classNames(
+                styles.ledgerConfirmAddressDescription,
+                ledgerModalConfirmDescriptionClassName
+              )}
+            >
+              {authTokenText?.data}
+            </div>
+          ) : null}
         </div>
       )}